Follow button - bugged

+++ GigHub/Controllers/Api/FollowingsController.cs
@@ -21,7 +21,7 @@ namespace GigHub.Controllers.Api
         {
             var userId = User.Identity.GetUserId();
 
-            if (_context.Followings.Any(f => f.FolloweeId == userId && f.FolloweeId == dto.FolloweeId))
+            if (_context.Followings.Any(f => f.FollowerId == userId && f.FolloweeId == dto.FolloweeId))
                 return BadRequest("Following already exists.");
 
             var following = new Following
@@ -43,10 +43,13 @@ namespace GigHub.Controllers.Api
             var following = _context.Followings
                 .Single(f => f.FollowerId == userId && f.FolloweeId == artistId);
 
+            if (following == null)
+                return NotFound();
+
             _context.Followings.Remove(following);
             _context.SaveChanges();
 
-            return Ok();
+            return Ok(artistId);
         }
     }
 }
